Implemented auto-refreshes for distros and other tweaks.
