+++ package.json
@@ -13,6 +13,7 @@
     "humanize-duration": "^3.18.0",
     "node-html-parser": "^1.1.15",
     "node-persist": "^3.0.4",
+    "node-schedule": "^1.3.2",
     "path": "^0.12.7",
     "simple-node-logger": "^18.12.22",
     "snekfetch": "^4.0.4"
